fix: Panic caused by a closed channel
In this commit I have fixed a panic that occurred when the kafka library sent messages in flight to a closed channel.

Currently, I have not found a way to prevent the library sending messages to the closed channel, so as a temporary fix I have decided not to close the channel.

NOTE: To avoid a deadlock the channel must remain a buffered channel.